Extract TimeStatistic
