+++ lib/domain/analysis/TimeLogAnalyser.ts
@@ -1,7 +1,8 @@
 import TimeLog from "../TimeLog";
-import { format, isBefore, max, min } from "date-fns";
+import { format } from "date-fns";
 import Workday from "../Workday";
 import { groupBy } from "lodash";
+import TimeStatistic from "./TimeStatistic";
 
 interface TimeStatistics {
   min: string;
@@ -26,7 +27,7 @@ export default class TimeLogAnalyser {
 
     Object.entries(groupedByWeekday).forEach(([weekday, workdays]) => {
       const startTimes = workdays.map(workday => workday.getStartTime());
-      const statistics = this.getTimeStatistics(startTimes);
+      const statistics = new TimeStatistic(startTimes).getTimeStatistics();
       result[weekday] = {
         min: format(statistics.min, "HH:mm:SS"),
         median: format(statistics.median, "HH:mm:SS"),
@@ -35,22 +36,5 @@ export default class TimeLogAnalyser {
     });
 
     return result;
-  }
-
-  private getTimeStatistics(times: Date[]) {
-    const minStartTime = min(times);
-    const medianStartTime = this.median(times);
-    const maxStartTime = max(times);
-    return {
-      min: minStartTime,
-      median: medianStartTime,
-      max: maxStartTime,
-    };
-  }
-
-  private median(times: Date[]) {
-    const sorted = times.sort((a, b) => isBefore(a, b) ? -1 : 1);
-    const middleIndex = Math.floor(times.length / 2);
-    return sorted[middleIndex];
   }
 }
